No need to add to extra applications in mix
extra applications is only useful for things that are part of the OTP or Elixir releases. Packages are tracked automatically from deps.

Also, in mix the preferred way to specify dependencies is with `~> major.minor` which should allow updating minor versions - those should be backwards compatible. Accidental updates are not a concern due to lockfiles.