correct the units shown for memory data
The units returned by `to_byte/1` are kilobytes (KB), megabytes (MB) and
gigabytes (GB) but the actual numeric value are for kibibyte (KiB), mebibyte(MiB)
and gibibyte(GiB). This is because the division is by 1024.

I found this problem when I was comparing the value agaist the output of
rebar:info and didn't match.

I did a similar patch a few years ago on OTP's observer [1].

[1] https://github.com/erlang/otp/pull/6063